added simple particle system script
changes the velocity of particles in the particle system. Currently using it on a particle system for water. Unforuntately, could not access force attributes for particles due to unity making them private/closed off. Still might be a way to manipulate the physics of things colliding with the water/submerged in water, but need to do some extra math and testing.